requirements.txt
numpy==1.26.4
opencv-python
mediapipe
tensorflow==2.16.1
ml-dtypes

pygame-ce==2.5.5
pymunk
pytmx

pytest
